refactor: PR review suggestions
