+++ agent/package.json
@@ -119,6 +119,11 @@
         "@elizaos/plugin-chainbase": "workspace:*",
         "@elizaos/plugin-dkg": "workspace:*",
         "@elizaos/plugin-email": "workspace:*",
+        "@elizaos/plugin-omniflix": "workspace:*",
+        "@elizaos/plugin-suno": "workspace:*",
+        "@elizaos/plugin-udio": "workspace:*",
+        "@elizaos/plugin-hyperbolic": "workspace:*",
+        "@elizaos/plugin-football": "workspace:*",
         "@elizaos/plugin-github": "workspace:*",
         "@elizaos/client-github": "workspace:*",
         "readline": "1.3.0",
@@ -133,4 +138,3 @@
         "tsup": "8.3.5"
     }
 }
-
